app/api/utils.py
import jwt
from datetime import datetime, timedelta, timezone
from functools import wraps
from flask import jsonify, request, current_app
from flask_login import current_user
from app.models import User


def generate_token(user):
    """Generate JWT token for user authentication."""
    payload = {
        'user_id': user.id,
        'username': user.username,
        'exp': datetime.now(timezone.utc) + timedelta(days=30),
        'iat': datetime.now(timezone.utc)
    }
    return jwt.encode(payload, current_app.config['SECRET_KEY'], algorithm='HS256')


def decode_token(token):
    """Decode JWT token and return user data."""
    try:
        payload = jwt.decode(token, current_app.config['SECRET_KEY'], algorithms=['HS256'])
        return payload
    except jwt.ExpiredSignatureError:
        return None
    except jwt.InvalidTokenError:
        return None


def token_required(f):
    """Decorator to require valid JWT token for API endpoints."""
    @wraps(f)
    def decorated(*args, **kwargs):
        token = request.headers.get('Authorization')
        
        if not token:
            return jsonify({'error': 'Token is missing'}), 401
            
        # Remove 'Bearer ' prefix if present
        if token.startswith('Bearer '):
            token = token[7:]
            
        payload = decode_token(token)
        if not payload:
            return jsonify({'error': 'Token is invalid or expired'}), 401
            
        # Get the user from the database
        user = User.query.get(payload['user_id'])
        if not user:
            return jsonify({'error': 'User not found'}), 401
            
        # Make user available in the request context
        request.current_user = user
        
        return f(*args, **kwargs)
    return decorated


def get_current_api_user():
    """Get the current user from the API request context."""
    return getattr(request, 'current_user', None)