requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Login==0.6.3
Flask-WTF==1.2.1
WTForms==3.1.1
Flask-Migrate==4.0.5
python-dotenv==1.0.0
psycopg2-binary==2.9.7
Werkzeug==2.3.7
email-validator
PyJWT==2.8.0
